Move tests outside project source folder
