+++ pytest.ini
@@ -1,5 +1,4 @@
 [pytest]
-testpaths = axes/tests
-python_files = tests.py test_*.py tests_*.py *_tests.py *_test.py
+testpaths = tests
 addopts = --cov axes --cov-config .coveragerc --cov-append --cov-report term-missing --cov-report=xml
-DJANGO_SETTINGS_MODULE = axes.tests.settings
+DJANGO_SETTINGS_MODULE = tests.settings
