+++ tests/settings.py
@@ -26,7 +26,7 @@ AUTHENTICATION_BACKENDS = [
 
 PASSWORD_HASHERS = ["django.contrib.auth.hashers.MD5PasswordHasher"]
 
-ROOT_URLCONF = "axes.tests.urls"
+ROOT_URLCONF = "tests.urls"
 
 INSTALLED_APPS = [
     "django.contrib.auth",
