+++ tests/test_attempts.py
@@ -6,8 +6,9 @@ from django.utils.timezone import now
 
 from axes.attempts import get_cool_off_threshold
 from axes.models import AccessAttempt
-from axes.tests.base import AxesTestCase
 from axes.utils import reset, reset_request
+
+from tests.base import AxesTestCase
 
 
 class GetCoolOffThresholdTestCase(AxesTestCase):
