+++ tests/test_logging.py
@@ -5,7 +5,8 @@ from django.urls import reverse
 
 from axes.apps import AppConfig
 from axes.models import AccessAttempt, AccessLog
-from axes.tests.base import AxesTestCase
+
+from tests.base import AxesTestCase
 
 
 @patch("axes.apps.AppConfig.initialized", False)
