+++ tests/test_management.py
@@ -5,7 +5,8 @@ from django.core.management import call_command
 from django.utils import timezone
 
 from axes.models import AccessAttempt, AccessLog
-from axes.tests.base import AxesTestCase
+
+from tests.base import AxesTestCase
 
 
 class ResetAccessLogsManagementCommandTestCase(AxesTestCase):
