+++ tests/test_models.py
@@ -5,7 +5,8 @@ from django.db.migrations.executor import MigrationExecutor
 from django.db.migrations.state import ProjectState
 
 from axes.models import AccessAttempt, AccessLog
-from axes.tests.base import AxesTestCase
+
+from tests.base import AxesTestCase
 
 
 class ModelsTestCase(AxesTestCase):
